aidlc-table-order/backend/pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "Table Order Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.0",
    "pydantic-settings",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "slowapi",
    "structlog",
    "python-json-logger",
    "boto3",
    "aioboto3",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "httpx",
    "factory-boy",
    "ruff",
    "mypy",
    "aiosqlite",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
